+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
+- **[UPDATE]** Normalize `TextArea`
 - **[UPDATE]** Update `ItemData` `data` prop to be nullable
+[...]
 
 # v48.0.0 (23/02/2021)